Gå over til den nye kontrakten for beregningsgrunnlag innhenting
